Stock.

+++ vs-sln/ECommerce.Stock.Domain/IStockEntryRepository.cs
@@ -2,7 +2,7 @@
 
 public interface IStockEntryRepository
 {
-	Task<IEnumerable<StockEntry>> ListByIdItem(Int32 itemId);
+	Task<IEnumerable<StockEntry>> ListByIdItem(Int64 itemId);
 	Task Save(StockEntry stockEntry);
 	Task Clean();
 }
