+++ vs-sln/ECommerce.Stock.Domain/StockEntry.cs
@@ -7,13 +7,13 @@ public enum Operation
 
 public sealed class StockEntry
 {
-    public Int32 ItemId { get; private set; }
+    public Int64 ItemId { get; private set; }
 
     public Operation Operation { get; private set; }
 
     public Int32 Quantity { get; private set; }
 
-    public StockEntry(Int32 itemId, Operation operation, Int32 quantity)
+    public StockEntry(Int64 itemId, Operation operation, Int32 quantity)
     {
         ItemId = itemId;
         Operation = operation;
